+++ src/index.js
@@ -1,2 +1,1 @@
-export { default as placePromos } from './placePromos'
 export { default as placementEngine } from './placementEngine'