+++ src/placePromos.js
@@ -1,49 +1,22 @@
-import UAParser from 'ua-parser-js'
 import chunkBy from 'fkit/dist/chunkBy'
-import compare from 'fkit/dist/compare'
-import compose from 'fkit/dist/compose'
+import compareBy from 'fkit/dist/compareBy'
 import copy from 'fkit/dist/copy'
 import filter from 'fkit/dist/filter'
 import get from 'fkit/dist/get'
 import head from 'fkit/dist/head'
 import map from 'fkit/dist/map'
 import pick from 'fkit/dist/pick'
+import pipe from 'fkit/dist/pipe'
 import sample from 'fkit/dist/sample'
 import sortBy from 'fkit/dist/sortBy'
-import toLower from 'fkit/dist/toLower'
-import toUpper from 'fkit/dist/toUpper'
 
 import runQuery from './runQuery'
-import { age } from './utils'
+import { xeqBy } from './utils'
 
 const PROMO_PROPERTIES = ['creativeId', 'promoId', 'slotId', 'groupId', 'campaignId']
 
 /**
- * Creates a new context to run a query. The functions and objects in the
- * context will be available to the query.
- *
- * @private
- */
-function createContext (user, window) {
-  const uaParser = new UAParser(get('navigator.userAgent', window))
-
-  return {
-    // Functions
-    age,
-    lower: toLower,
-    upper: toUpper,
-
-    // Objects
-    browser: uaParser.getBrowser(),
-    device: uaParser.getDevice(),
-    os: uaParser.getOS(),
-    user,
-    window
-  }
-}
-
-/**
- * Filter the promos that have satisfied constraints.
+ * Filters the promos which satisfy their constraints in the given context.
  *
  * @private
  */
@@ -60,41 +33,33 @@ function filterPromos (context) {
 }
 
 /**
- * Filters promos based on which promo constraints are satisfied by the client
- * state.
- *
- * Placement rules:
+ * Places the promos based on the following rules:
  *
  *   - Promos with satisfied constraints will be placed, otherwise they are skipped.
  *   - If a group contains two or more promos, then one will be picked at random.
  *   - Promos without a group are picked independently of each other.
  *
+ * @params {Object} context The context object.
  * @params {Array} promos The list of promos to place.
- * @params {Object} user The user object.
- * @params {Window} window The window object.
  * @returns {Array} The list of placed promos.
  */
-export default function placePromos (promos, user, window) {
-  // Create a new context.
-  const context = createContext(user, window)
+export default function placePromos (context, promos) {
+  return pipe([
+    // Filter the promos.
+    filterPromos(context),
 
-  // Filter the promos.
-  const f = filterPromos(context)
+    // Sort the promos by groupId.
+    sortBy(compareBy(get('groupId'))),
 
-  // Sort the promos by group.
-  const g = sortBy((a, b) => compare(a.groupId, b.groupId))
+    // Chunk the promos by groupId. Promos with matching groupIds will be placed
+    // in the same chunk. Promos without a groupId will be placed in their own
+    // chunk.
+    chunkBy(xeqBy(get('groupId'))),
 
-  // Chunk the promos by group. Promos without a group are considered to be in
-  // a group of their own.
-  const h = chunkBy((a, b) => a.groupId && b.groupId && a.groupId === b.groupId)
-
-  // Choose one random promo from each group.
-  //
-  // Only one promo from each group may be placed at the same time. This allows
-  // us to randomly cycle through multiple variants of a promo.
-  const i = map(group => head(sample(1, group)))
-
-  // Compose the transform functions and apply the promos to get the list of
-  // placed promos.
-  return compose(i, h, g, f)(promos)
+    // Choose one random promo from each group.
+    //
+    // Only one promo from each group may be placed at the same time. This allows
+    // us to randomly cycle through multiple variants of a promo.
+    map(group => head(sample(1, group)))
+  ])(promos)
 }