+++ src/userState.js
@@ -15,9 +15,10 @@ export function get (storage) {
 /**
  * Stores the given user state.
  *
- * @param storage The storage object.
+ * @param {Storage} storage The storage object.
  * @param {Object} user The user state.
  */
 export function set (storage, user) {
   storage.setItem('user', JSON.stringify(user))
+  return user
 }